+++ src/components/grid/grid.component.ts
@@ -5,7 +5,7 @@ import {Observable, BehaviorSubject} from 'rxjs';
 import {GridPluginInstances, Grid, GridFunction} from "./grid.interface";
 import {GRID_PLUGIN_INSTANCES} from './types';
 import {GridOptions, PluginDescription, GridPlugin} from "../../misc";
-import {GRID_OPTIONS, PAGING_TYPE, DATA_LOADER_TYPE, CONTENT_RENDERER_TYPE, METADATA_SELECTOR_TYPE, NO_DATA_RENDERER_TYPE, ROW_SELECTOR_TYPE, PAGING_INITIALIZER_TYPE, GRID_INITIALIZER_TYPE} from "../../misc/types";
+import {GRID_OPTIONS, PAGING_TYPE, DATA_LOADER_TYPE, CONTENT_RENDERER_TYPE, METADATA_SELECTOR_TYPE, NO_DATA_RENDERER_TYPE, ROW_SELECTOR_TYPE, GRID_INITIALIZER_TYPE} from "../../misc/types";
 import {PagingPosition} from "../../misc/enums";
 import {Paging} from "../../plugins/paging";
 import {BasicPagingComponent} from "../../plugins/paging/components";
@@ -27,9 +27,6 @@ import {NO_DATA_RENDERER} from "../../plugins/noDataRenderer/types";
 import {RowSelector} from "../../plugins/rowSelector";
 import {BasicRowSelectorComponent} from "../../plugins/rowSelector/components";
 import {ROW_SELECTOR} from "../../plugins/rowSelector/types";
-import {PagingInitializer} from "../../plugins/pagingInitializer";
-import {NoPagingInitializerComponent} from "../../plugins/pagingInitializer/components";
-import {PAGING_INITIALIZER} from "../../plugins/pagingInitializer/types";
 import {GridInitializer} from "../../plugins/gridInitializer";
 import {NoGridInitializerComponent} from "../../plugins/gridInitializer/components";
 import {GRID_INITIALIZER} from "../../plugins/gridInitializer/types";
@@ -69,10 +66,6 @@ const defaultOptions: GridOptions =
         rowSelector: <PluginDescription<BasicRowSelectorComponent<any, any, any>>>
         {
             type: forwardRef(() => BasicRowSelectorComponent)
-        },
-        pagingInitializer: <PluginDescription<NoPagingInitializerComponent>>
-        {
-            type: forwardRef(() => NoPagingInitializerComponent)
         },
         gridInitializer: <PluginDescription<NoGridInitializerComponent>>
         {
@@ -161,7 +154,6 @@ export class GridComponent implements OnInit, AfterViewInit, Grid
                 @Inject(GRID_PLUGIN_INSTANCES) private _pluginInstances: GridPluginInstances,
                 @Inject(GRID_OPTIONS) @Optional() options?: GridOptions,
                 @Inject(PAGING_TYPE) @Optional() pagingType?: Type<Paging>,
-                @Inject(PAGING_INITIALIZER_TYPE) @Optional() pagingInitializerType?: Type<PagingInitializer>,
                 @Inject(GRID_INITIALIZER_TYPE) @Optional() gridInitializerType?: Type<GridInitializer>,
                 @Inject(DATA_LOADER_TYPE) @Optional() dataLoaderType?: Type<DataLoader<any>>,
                 @Inject(CONTENT_RENDERER_TYPE) @Optional() contentRendererType?: Type<ContentRenderer<any>>,
@@ -184,16 +176,6 @@ export class GridComponent implements OnInit, AfterViewInit, Grid
             }
 
             opts.plugins.paging.type = pagingType;
-        }
-
-        if(pagingInitializerType)
-        {
-            if(!opts.plugins.pagingInitializer)
-            {
-                opts.plugins.pagingInitializer = {};
-            }
-
-            opts.plugins.pagingInitializer.type = pagingInitializerType;
         }
 
         if(gridInitializerType)
@@ -308,33 +290,6 @@ export class GridComponent implements OnInit, AfterViewInit, Grid
         if(this._gridOptions.plugins && this._gridOptions.plugins.paging && this._gridOptions.plugins.paging.instanceCallback)
         {
             this._gridOptions.plugins.paging.instanceCallback(paging);
-        }
-    }
-
-    /**
-     * Sets paging initializer component
-     * @param pagingInitializer - Created paging initializer that is used
-     * @internal
-     */
-    public setPagingInitializerComponent(pagingInitializer: PagingInitializer)
-    {
-        if(!pagingInitializer)
-        {
-            return;
-        }
-
-        this._pluginInstances[PAGING_INITIALIZER] = pagingInitializer;
-
-        if(this._gridOptions.plugins && this._gridOptions.plugins.pagingInitializer && this._gridOptions.plugins.pagingInitializer.options)
-        {
-            pagingInitializer.options = this._gridOptions.plugins.pagingInitializer.options;
-        }
-
-        pagingInitializer.initOptions();
-        
-        if(this._gridOptions.plugins && this._gridOptions.plugins.pagingInitializer && this._gridOptions.plugins.pagingInitializer.instanceCallback)
-        {
-            this._gridOptions.plugins.pagingInitializer.instanceCallback(pagingInitializer);
         }
     }
 
@@ -511,7 +466,6 @@ export class GridComponent implements OnInit, AfterViewInit, Grid
     {
         this._pluginInstances[ROW_SELECTOR].initialize();
         this._pluginInstances[METADATA_SELECTOR].initialize();
-        this._pluginInstances[PAGING_INITIALIZER].initialize();
         this._pluginInstances[GRID_INITIALIZER].initialize();
         this._pluginInstances[PAGING].initialize();
         this._pluginInstances[CONTENT_RENDERER].initialize();
@@ -547,28 +501,6 @@ export class GridComponent implements OnInit, AfterViewInit, Grid
                     }
 
                     this._pluginInstances[PAGING].initOptions();
-                }
-            }
-
-            if(this._gridOptions.plugins.pagingInitializer)
-            {
-                this._gridOptions.plugins.pagingInitializer.type = resolveForwardRef(this._gridOptions.plugins.pagingInitializer.type);
-
-                if(this._gridOptions.plugins.pagingInitializer.instance &&
-                   this._gridOptions.plugins.pagingInitializer.instance != this._pluginInstances[PAGING_INITIALIZER])
-                {
-                    this._pluginInstances[PAGING_INITIALIZER] = this._gridOptions.plugins.pagingInitializer.instance;
-                    this._gridOptions.plugins.pagingInitializer.instance.gridPlugins = this._pluginInstances;
-                }
-
-                if(this._pluginInstances[PAGING_INITIALIZER])
-                {
-                    if(this._gridOptions.plugins && this._gridOptions.plugins.pagingInitializer && this._gridOptions.plugins.pagingInitializer.options)
-                    {
-                        this._pluginInstances[PAGING_INITIALIZER].options = this._gridOptions.plugins.pagingInitializer.options;
-                    }
-
-                    this._pluginInstances[PAGING_INITIALIZER].initOptions();
                 }
             }