+++ src/plugins/pagingInitializer/components.ts
@@ -1,3 +1,0 @@
-export * from './no/noPagingInitializer.component';
-export * from './query/queryPagingInitializer.component';
-export * from './queryCookie/queryCookiePagingInitializer.component';