+++ src/plugins/pagingInitializer/index.ts
@@ -1,4 +1,0 @@
-export * from './pagingInitializer.interface';
-export * from './no/noPagingInitializer.interface';
-export * from './query/queryPagingInitializer.interface';
-export * from './queryCookie/queryCookiePagingInitializer.interface';